Replace awesome_print with amazing_print
The `awesome_print` gem has not been maintained for several years.

There is now a successor gem named `amazing_print`, which implements the
same API but is more actively maintained.

Updating a project to use this commit will require some snapshots to be
regenerated because `amazing_print` produces a slightly different result
from that of the `awesome_print` gem. For example, it can serialize
hashes using the newer `key: value` syntax instead of hash rockets.